requirements.txt
pywebview
pyinstaller
matplotlib
psutil
debugpy
nuitka